Update depend files, etc (#499)
* Add missing depend files

* Remove needless size_t to uint32_t cast

+++ ext/bigdecimal/bigdecimal.c
@@ -4949,7 +4949,7 @@ VpMult(Real *c, Real *a, Real *b)
     if (!AddExponent(c, b->exponent)) return 0;
 
     if (b->Prec >= NTT_MULTIPLICATION_THRESHOLD) {
-        ntt_multiply((uint32_t)a->Prec, (uint32_t)b->Prec, a->frac, b->frac, c->frac);
+        ntt_multiply(a->Prec, b->Prec, a->frac, b->frac, c->frac);
         c->Prec = a->Prec + b->Prec;
         goto Cleanup;
     }
